+++ units/src/fee_rate/mod.rs
@@ -10,9 +10,6 @@ use core::{fmt, ops};
 #[cfg(feature = "arbitrary")]
 use arbitrary::{Arbitrary, Unstructured};
 
-use crate::amount::Amount;
-use crate::weight::Weight;
-
 /// Represents fee rate.
 ///
 /// This is an integer newtype representing fee rate in `sat/kwu`. It provides protection against mixing
@@ -101,26 +98,6 @@ impl FeeRate {
         }
     }
 
-    /// Checked weight multiplication.
-    ///
-    /// Computes the absolute fee amount for a given [`Weight`] at this fee rate.
-    /// When the resulting fee is a non-integer amount, the amount is rounded up,
-    /// ensuring that the transaction fee is enough instead of falling short if
-    /// rounded down.
-    ///
-    /// [`None`] is returned if an overflow occurred.
-    #[must_use]
-    pub const fn checked_mul_by_weight(self, weight: Weight) -> Option<Amount> {
-        // No `?` operator in const context.
-        match self.to_sat_per_kwu().checked_mul(weight.to_wu()) {
-            Some(mul_res) => match mul_res.checked_add(999) {
-                Some(add_res) => Some(Amount::from_sat(add_res / 1000)),
-                None => None,
-            },
-            None => None,
-        }
-    }
-
     /// Checked addition.
     ///
     /// Computes `self + rhs` returning [`None`] if overflow occurred.
@@ -143,23 +120,6 @@ impl FeeRate {
             Some(res) => Some(Self(res)),
             None => None,
         }
-    }
-
-    /// Calculates the fee by multiplying this fee rate by weight, in weight units, returning [`None`]
-    /// if an overflow occurred.
-    ///
-    /// This is equivalent to `Self::checked_mul_by_weight()`.
-    #[must_use]
-    pub fn fee_wu(self, weight: Weight) -> Option<Amount> { self.checked_mul_by_weight(weight) }
-
-    /// Calculates the fee by multiplying this fee rate by weight, in virtual bytes, returning [`None`]
-    /// if an overflow occurred.
-    ///
-    /// This is equivalent to converting `vb` to [`Weight`] using [`Weight::from_vb`] and then calling
-    /// `Self::fee_wu(weight)`.
-    #[must_use]
-    pub fn fee_vb(self, vb: u64) -> Option<Amount> {
-        Weight::from_vb(vb).and_then(|w| self.fee_wu(w))
     }
 }
 
@@ -193,33 +153,6 @@ impl ops::Sub for FeeRate {
 }
 crate::internal_macros::impl_sub_for_references!(FeeRate);
 crate::internal_macros::impl_sub_assign!(FeeRate);
-
-/// Computes the ceiling so that the fee computation is conservative.
-impl ops::Mul<FeeRate> for Weight {
-    type Output = Amount;
-
-    fn mul(self, rhs: FeeRate) -> Self::Output {
-        Amount::from_sat((rhs.to_sat_per_kwu() * self.to_wu() + 999) / 1000)
-    }
-}
-
-impl ops::Mul<Weight> for FeeRate {
-    type Output = Amount;
-
-    fn mul(self, rhs: Weight) -> Self::Output { rhs * self }
-}
-
-impl ops::Div<Weight> for Amount {
-    type Output = FeeRate;
-
-    /// Truncating integer division.
-    ///
-    /// This is likely the wrong thing for a user dividing an amount by a weight. Consider using
-    /// `checked_div_by_weight` instead.
-    fn div(self, rhs: Weight) -> Self::Output {
-        FeeRate::from_sat_per_kwu(self.to_sat() * 1000 / rhs.to_wu())
-    }
-}
 
 impl core::iter::Sum for FeeRate {
     fn sum<I>(iter: I) -> Self
@@ -292,16 +225,6 @@ mod tests {
     }
 
     #[test]
-    #[allow(clippy::op_ref)]
-    fn multiply() {
-        let two = FeeRate::from_sat_per_vb(2).unwrap();
-        let three = Weight::from_vb(3).unwrap();
-        let six = Amount::from_sat(6);
-
-        assert_eq!(two * three, six);
-    }
-
-    #[test]
     fn add_assign() {
         let mut f = FeeRate(1);
         f += FeeRate(2);
@@ -321,12 +244,6 @@ mod tests {
         let mut f = FeeRate(3);
         f -= &FeeRate(2);
         assert_eq!(f, FeeRate(1));
-    }
-
-    #[test]
-    fn fee_rate_div_by_weight() {
-        let fee_rate = Amount::from_sat(329) / Weight::from_wu(381);
-        assert_eq!(fee_rate, FeeRate::from_sat_per_kwu(863));
     }
 
     #[test]
@@ -400,50 +317,6 @@ mod tests {
 
         let fee_rate = FeeRate(10).checked_mul(u64::MAX);
         assert!(fee_rate.is_none());
-    }
-
-    #[test]
-    fn fee_wu() {
-        let fee_overflow = FeeRate::from_sat_per_kwu(10).fee_wu(Weight::MAX);
-        assert!(fee_overflow.is_none());
-
-        let fee_rate = FeeRate::from_sat_per_vb(2).unwrap();
-        let weight = Weight::from_vb(3).unwrap();
-        assert_eq!(fee_rate.fee_wu(weight).unwrap(), Amount::from_sat(6));
-    }
-
-    #[test]
-    fn fee_vb() {
-        let fee_overflow = FeeRate::from_sat_per_kwu(10).fee_vb(Weight::MAX.to_wu());
-        assert!(fee_overflow.is_none());
-
-        let fee_rate = FeeRate::from_sat_per_vb(2).unwrap();
-        assert_eq!(fee_rate.fee_vb(3).unwrap(), Amount::from_sat(6));
-    }
-
-    #[test]
-    fn checked_weight_mul() {
-        let weight = Weight::from_vb(10).unwrap();
-        let fee: Amount = FeeRate::from_sat_per_vb(10)
-            .unwrap()
-            .checked_mul_by_weight(weight)
-            .expect("expected Amount");
-        assert_eq!(Amount::from_sat(100), fee);
-
-        let fee = FeeRate::from_sat_per_kwu(10).checked_mul_by_weight(Weight::MAX);
-        assert!(fee.is_none());
-
-        let weight = Weight::from_vb(3).unwrap();
-        let fee_rate = FeeRate::from_sat_per_vb(3).unwrap();
-        let fee = fee_rate.checked_mul_by_weight(weight).unwrap();
-        assert_eq!(Amount::from_sat(9), fee);
-
-        let weight = Weight::from_wu(381);
-        let fee_rate = FeeRate::from_sat_per_kwu(864);
-        let fee = fee_rate.checked_mul_by_weight(weight).unwrap();
-        // 381 * 0.864 yields 329.18.
-        // The result is then rounded up to 330.
-        assert_eq!(fee, Amount::from_sat(330));
     }
 
     #[test]
